+++ src/test/java/kr/KWGraduate/BookPharmacy/repository/AnswerRepositoryTest.java
@@ -1,9 +1,12 @@
 package kr.KWGraduate.BookPharmacy.repository;
 
-import kr.KWGraduate.BookPharmacy.entity.Answer;
-import kr.KWGraduate.BookPharmacy.entity.Board;
-import kr.KWGraduate.BookPharmacy.entity.Client;
-import kr.KWGraduate.BookPharmacy.entity.Keyword;
+import kr.KWGraduate.BookPharmacy.domain.answer.domain.Answer;
+import kr.KWGraduate.BookPharmacy.domain.answer.repository.AnswerRepository;
+import kr.KWGraduate.BookPharmacy.domain.board.domain.Board;
+import kr.KWGraduate.BookPharmacy.domain.board.repository.BoardRepository;
+import kr.KWGraduate.BookPharmacy.domain.client.domain.Client;
+import kr.KWGraduate.BookPharmacy.domain.client.repository.ClientRepository;
+import kr.KWGraduate.BookPharmacy.domain.keyword.domain.Keyword;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
